Refactor DashboardActivity with DashboardViewModel
